--- instrumentation/runtime-metrics/javaagent/build.gradle.kts
@@ -1,9 +1,0 @@
-plugins {
-  id("otel.javaagent-instrumentation")
-}
-
-dependencies {
-  implementation(project(":instrumentation:runtime-metrics:library"))
-
-  compileOnly("io.opentelemetry:opentelemetry-sdk-extension-autoconfigure")
-}
