--- instrumentation/runtime-metrics/javaagent/src/main/java/io/opentelemetry/instrumentation/javaagent/runtimemetrics/RuntimeMetricsInstaller.java
@@ -1,43 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package io.opentelemetry.instrumentation.javaagent.runtimemetrics;
-
-import com.google.auto.service.AutoService;
-import io.opentelemetry.api.GlobalOpenTelemetry;
-import io.opentelemetry.api.OpenTelemetry;
-import io.opentelemetry.instrumentation.runtimemetrics.BufferPools;
-import io.opentelemetry.instrumentation.runtimemetrics.Classes;
-import io.opentelemetry.instrumentation.runtimemetrics.Cpu;
-import io.opentelemetry.instrumentation.runtimemetrics.GarbageCollector;
-import io.opentelemetry.instrumentation.runtimemetrics.MemoryPools;
-import io.opentelemetry.instrumentation.runtimemetrics.Threads;
-import io.opentelemetry.javaagent.extension.AgentListener;
-import io.opentelemetry.sdk.autoconfigure.AutoConfiguredOpenTelemetrySdk;
-import io.opentelemetry.sdk.autoconfigure.spi.ConfigProperties;
-
-/** An {@link AgentListener} that enables runtime metrics during agent startup. */
-@AutoService(AgentListener.class)
-public class RuntimeMetricsInstaller implements AgentListener {
-
-  @Override
-  public void afterAgent(AutoConfiguredOpenTelemetrySdk autoConfiguredSdk) {
-    ConfigProperties config = autoConfiguredSdk.getConfig();
-
-    boolean defaultEnabled = config.getBoolean("otel.instrumentation.common.default-enabled", true);
-    if (!config.getBoolean("otel.instrumentation.runtime-metrics.enabled", defaultEnabled)) {
-      return;
-    }
-
-    OpenTelemetry openTelemetry = GlobalOpenTelemetry.get();
-
-    BufferPools.registerObservers(openTelemetry);
-    Classes.registerObservers(openTelemetry);
-    Cpu.registerObservers(openTelemetry);
-    MemoryPools.registerObservers(openTelemetry);
-    Threads.registerObservers(openTelemetry);
-    GarbageCollector.registerObservers(openTelemetry);
-  }
-}
