+++ instrumentation/runtime-telemetry-jmx/javaagent/build.gradle.kts
@@ -1,0 +1,9 @@
+plugins {
+  id("otel.javaagent-instrumentation")
+}
+
+dependencies {
+  implementation(project(":instrumentation:runtime-telemetry-jmx:library"))
+
+  compileOnly("io.opentelemetry:opentelemetry-sdk-extension-autoconfigure")
+}
