+++ instrumentation/runtime-telemetry-jmx/library/src/main/java/io/opentelemetry/instrumentation/runtimetelemetryjmx/JmxRuntimeMetricsUtil.java
@@ -1,0 +1,31 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package io.opentelemetry.instrumentation.runtimetelemetryjmx;
+
+import io.opentelemetry.api.OpenTelemetry;
+import io.opentelemetry.api.metrics.Meter;
+import io.opentelemetry.api.metrics.MeterBuilder;
+import io.opentelemetry.instrumentation.api.internal.EmbeddedInstrumentationProperties;
+import javax.annotation.Nullable;
+
+class JmxRuntimeMetricsUtil {
+
+  private static final String INSTRUMENTATION_NAME = "io.opentelemetry.runtime-telemetry-jmx";
+
+  @Nullable
+  private static final String INSTRUMENTATION_VERSION =
+      EmbeddedInstrumentationProperties.findVersion(INSTRUMENTATION_NAME);
+
+  static Meter getMeter(OpenTelemetry openTelemetry) {
+    MeterBuilder meterBuilder = openTelemetry.meterBuilder(INSTRUMENTATION_NAME);
+    if (INSTRUMENTATION_VERSION != null) {
+      meterBuilder.setInstrumentationVersion(INSTRUMENTATION_VERSION);
+    }
+    return meterBuilder.build();
+  }
+
+  private JmxRuntimeMetricsUtil() {}
+}
