+++ instrumentation/runtime-telemetry-jmx/library/src/test/java/io/opentelemetry/instrumentation/runtimetelemetryjmx/ScopeUtil.java
@@ -1,0 +1,25 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package io.opentelemetry.instrumentation.runtimetelemetryjmx;
+
+import io.opentelemetry.instrumentation.api.internal.EmbeddedInstrumentationProperties;
+import io.opentelemetry.sdk.common.InstrumentationScopeInfo;
+import java.util.Optional;
+
+class ScopeUtil {
+
+  static final InstrumentationScopeInfo EXPECTED_SCOPE =
+      InstrumentationScopeInfo.builder("io.opentelemetry.runtime-telemetry-jmx")
+          .setVersion(
+              Optional.ofNullable(
+                      EmbeddedInstrumentationProperties.findVersion(
+                          "io.opentelemetry.runtime-telemetry-jmx"))
+                  .orElseThrow(
+                      () -> new IllegalStateException("Unable to find instrumentation version")))
+          .build();
+
+  private ScopeUtil() {}
+}
