+++ settings.gradle.kts
@@ -433,8 +433,8 @@ hideFromDependabot(":instrumentation:rocketmq:rocketmq-client:rocketmq-client-4.
 hideFromDependabot(":instrumentation:rocketmq:rocketmq-client:rocketmq-client-4.8:testing")
 hideFromDependabot(":instrumentation:rocketmq:rocketmq-client:rocketmq-client-5.0:javaagent")
 hideFromDependabot(":instrumentation:rocketmq:rocketmq-client:rocketmq-client-5.0:testing")
-hideFromDependabot(":instrumentation:runtime-metrics:javaagent")
-hideFromDependabot(":instrumentation:runtime-metrics:library")
+hideFromDependabot(":instrumentation:runtime-telemetry-jmx:javaagent")
+hideFromDependabot(":instrumentation:runtime-telemetry-jmx:library")
 hideFromDependabot(":instrumentation:runtime-telemetry-jfr:javaagent")
 hideFromDependabot(":instrumentation:runtime-telemetry-jfr:library")
 hideFromDependabot(":instrumentation:rxjava:rxjava-1.0:library")
